requirements.txt
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
Flask==2.2.1
Flask-Bcrypt==1.0.1
flask-msearch==0.2.9.2
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
SQLAlchemy==1.4.39
Werkzeug==2.2.2
WTForms==3.0.1
